Module der::asn1

Module containing all of the various ASN.1 built-in types supported by this library.

Structs

AnyRef

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

BitStringIter

Iterator over the bits of a [BitString].

BitStringRef

ASN.1 BIT STRING type.

ContextSpecific

Context-specific field which wraps an owned inner value.

ContextSpecificRef

Context-specific field reference.

GeneralizedTime

ASN.1 GeneralizedTime type.

Ia5StringRef

ASN.1 IA5String type.

Null

ASN.1 NULL type.

ObjectIdentifieroid

Object identifier (OID).

OctetStringRef

ASN.1 OCTET STRING type: borrowed form.

PrintableStringRef

ASN.1 PrintableString type.

SequenceOf

ASN.1 SEQUENCE OF backed by an array.

SequenceOfIter

Iterator over the elements of an SequenceOf.

SequenceRef

The SequenceRef type provides raw access to the octets which comprise a DER-encoded SEQUENCE.

SetOf

ASN.1 SET OF backed by an array.

SetOfIter

Iterator over the elements of an SetOf.

TeletexStringRef

ASN.1 TeletexString type.

UIntRef

“Big” unsigned ASN.1 INTEGER type.

UtcTime

ASN.1 UTCTime type.

Utf8StringRef

ASN.1 UTF8String type.

VideotexStringRef

ASN.1 VideotexString type.

Traits

Choice

ASN.1 CHOICE denotes a union of one or more possible alternatives.

Sequence

ASN.1 SEQUENCE trait.