der::asn1::integer

Module int

Support for encoding signed integers

Re-exports

pub use allocating::Int;

Modules

allocating 🔒

Macros

impl_encoding_traits 🔒

Structs

IntRef

Signed arbitrary precision ASN.1 INTEGER reference type.

Functions

decode_to_array 🔒

Decode an signed integer of the specified size.

encode_bytes 🔒

Encode the given big endian bytes representing an integer as ASN.1 DER.

negative_encoded_len 🔒

Get the encoded length for the given negative integer serialized as bytes.

strip_leading_ones 🔒

Strip the leading all-ones bytes from the given byte slice.

validate_canonical 🔒

Ensure INTEGER is canonically encoded.