der::asn1::integer

Module uint

Unsigned integer decoders/encoders.

Re-exports

pub use allocating::Uint;

Modules

allocating 🔒

Macros

impl_encoding_traits 🔒

Structs

UintRef

Unsigned arbitrary precision ASN.1 INTEGER reference type.

Functions

decode_to_array 🔒

Decode an unsigned integer into a byte array of the requested size containing a big endian integer.

decode_to_slice 🔒

Decode an unsigned integer into a big endian byte slice with all leading zeroes removed.

encode_bytes 🔒

Encode the given big endian bytes representing an integer as ASN.1 DER.

encoded_len 🔒

Get the encoded length for the given unsigned integer serialized as bytes.

needs_leading_zero 🔒

Does the given integer need a leading zero?

strip_leading_zeroes 🔒

Strip the leading zeroes from the given byte slice