der::asn1

Module set_of

ASN.1 SET OF support.

Ordering Notes

Some DER serializer implementations fail to properly sort elements of a SET OF. This is technically non-canonical, but occurs frequently enough that most DER decoders tolerate it.

When decoding with EncodingRules::Der, this implementation sorts the elements of SET OF at decode-time to ensure reserializations are canonical.

Modules

allocating 🔒

Structs

SetOfIter

Iterator over the elements of an [SetOf].

SetOfRef

ASN.1 SET OF with a reference to an array.

SetOfRefIter

Iterator over the elements of an SetOfRef.

SetOfVec

ASN.1 SET OF backed by a Vec.

Enums

InnerIterRef 🔒

InnerRef 🔒

Functions

check_der_ordering 🔒

Ensure set elements are lexicographically ordered using DerOrd.

der_sort 🔒

Sort a mut slice according to its DerOrd, returning any errors which might occur during the comparison.