der::asn1

Struct Any

pub struct Any { /* private fields */ }

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

This type provides the same functionality as AnyRef but owns the backing data.

Implementations

impl Any

pub fn new(tag: Tag, bytes: impl Into<Box<[u8]>>) -> Result<Self, Error>

Create a new Any from the provided Tag and DER bytes.

pub fn value(&self) -> &[u8]

Allow access to value

pub fn header(&self) -> Header

Returns Tag and Length of self.

pub fn decode_as<'a, T>(&'a self) -> Result<T, <T as DecodeValue<'a>>::Error>
where T: Choice<'a> + DecodeValue<'a>,

Attempt to decode this Any type into the inner value.

pub fn decode_as_encoding<'a, T>( &'a self, encoding: EncodingRules, ) -> Result<T, <T as DecodeValue<'a>>::Error>
where T: Choice<'a> + DecodeValue<'a>,

Attempt to decode this Any type into the inner value with the given encoding rules.

pub fn encode_from<T>(msg: &T) -> Result<Self, Error>
where T: Tagged + EncodeValue,

Encode the provided type as an Any value.

pub fn sequence<'a, F, T, E>(&'a self, f: F) -> Result<T, E>
where F: FnOnce(&mut SliceReader<'a>) -> Result<T, E>, E: From<Error>,

Attempt to decode this value an ASN.1 SEQUENCE, creating a new nested reader and calling the provided argument with it.

pub fn null() -> Self

Any representation of the ASN.1 NULL type.

pub fn to_ref(&self) -> AnyRef<'_>

Create a new AnyRef from the provided Any owned tag and bytes.

impl Any

pub fn is_null(&self) -> bool

Is this value an ASN.1 NULL value?