Struct der::asn1::ContextSpecific

pub struct ContextSpecific<'a> {
    pub tag_number: TagNumber,
    pub value: Any<'a>,
}

Context-specific field.

This type encodes a field which is specific to a particular context, and is identified by a TagNumber.

Any context-specific field can be decoded/encoded with this type. The intended use is to dynamically dispatch off of the context-specific tag number when decoding, which allows support for extensions, which are denoted in an ASN.1 schema using the ... ellipsis extension marker.

Fields
tag_number: TagNumber

Context-specific tag number sans the leading 0b10000000 class identifier bit and 0b100000 constructed flag.

value: Any<'a>

Value of the field.