Struct der::asn1::ContextSpecificRef

pub struct ContextSpecificRef<'a, T> {
    pub tag_number: TagNumber,
    pub tag_mode: TagMode,
    pub value: &'a T,
}

Context-specific field reference.

This type encodes a field which is specific to a particular context and is identified by a TagNumber.

Fields
tag_number: TagNumber

Context-specific tag number sans the leading 0b10000000 class identifier bit and 0b100000 constructed flag.

tag_mode: TagMode

Tag mode: EXPLICIT VS IMPLICIT.

value: &'a T

Value of the field.