Struct der::asn1::Sequence

pub struct Sequence<'a> { /* private fields */ }

ASN.1 SEQUENCE type.

Implementations

impl<'a> Sequence<'a>

pub fn as_bytes(&self) -> &'a [u8]

Borrow the inner byte sequence.

pub fn decode_nested<F, T>(&self, f: F) -> Result<T>
where F: FnOnce(&mut Decoder<'a>) -> Result<T>,

Decode values nested within a sequence, creating a new Decoder for the data contained in the sequenceโs body and passing it to the provided FnOnce.

pub fn iter<T: Decodable<'a>>(&self) -> SequenceIter<'a, T>

Iterate over the values in a heterogenously typed sequence.