der::asn1

Struct SequenceRef

pub struct SequenceRef { /* private fields */ }

The SequenceRef type provides raw access to the octets which comprise a DER-encoded SEQUENCE.

This is a zero-copy reference type which borrows from the input data.

Implementations

impl SequenceRef

pub fn new(slice: &[u8]) -> Result<&Self>

Create a new ASN.1 SEQUENCE from a byte slice.

Errors

Returns Error in the event slice is too long.

pub fn as_bytes(&self) -> &[u8]

Borrow the inner byte slice.