Struct der::asn1::UIntBytes

pub struct UIntBytes<'a> { /* private fields */ }

“Big” unsigned ASN.1 INTEGER type.

Provides direct access to the underlying big endian bytes which comprise an unsigned integer value.

Intended for use cases like very large integers that are used in cryptographic applications (e.g. keys, signatures).

Implementations

impl<'a> UIntBytes<'a>

pub fn new(bytes: &'a [u8]) -> Result<Self>

Create a new UIntBytes from a byte slice.

pub fn as_bytes(&self) -> &'a [u8]

Borrow the inner byte slice which contains the least significant bytes of a big endian integer value with all leading zeros stripped.

pub fn len(&self) -> Length

Get the length of this UIntBytes in bytes.

pub fn is_empty(&self) -> bool

Is the inner byte slice empty?