Struct der::asn1::Utf8String

pub struct Utf8String<'a> { /* private fields */ }

ASN.1 UTF8String type.

Supports the full UTF-8 encoding.

Note that the Decodable and Encodable traits are impl’d for Rust’s str primitive, which decodes/encodes as a Utf8String.

You are free to use str instead of this type, however it’s still provided for explicitness in cases where it might be ambiguous with other ASN.1 string encodings such as PrintableString.

Implementations

impl<'a> Utf8String<'a>

pub fn new<T>(input: &'a T) -> Result<Self>
where T: AsRef<[u8]> + ?Sized,

Create a new ASN.1 UTF8String.

pub fn as_str(&self) -> &'a str

Borrow the string as a str.

pub fn as_bytes(&self) -> &'a [u8]

Borrow the string as bytes.

pub fn len(&self) -> Length

Get the length of the inner byte slice.

pub fn is_empty(&self) -> bool

Is the inner string empty?