Struct der::asn1::Utf8StringRef

pub struct Utf8StringRef<'a> { /* private fields */ }

ASN.1 UTF8String type.

Supports the full UTF-8 encoding.

Note that the Decode and Encode traits are impl’d for Rust’s str primitive, which decodes/encodes as a Utf8StringRef.

You are free to use str instead of this type, however it’s still provided for explicitness in cases where it might be ambiguous with other ASN.1 string encodings such as PrintableStringRef.

This is a zero-copy reference type which borrows from the input data.

Implementations

impl<'a> Utf8StringRef<'a>

pub fn new<T>(input: &'a T) -> Result<Self>
where T: AsRef<[u8]> + ?Sized,

Create a new ASN.1 UTF8String.