der::bytes

Struct BytesRef

#[repr(transparent)]
pub(crate) struct BytesRef([u8]);

Byte slice newtype which respects the Length::MAX limit.

Tuple Fields
0: [u8]
Implementations

impl BytesRef

pub const EMPTY: &'static Self

Constant value representing an empty byte slice.

pub const fn new(slice: &[u8]) -> Result<&Self>

Create a new BytesRef, ensuring that the provided slice value is shorter than Length::MAX.

pub(crate) const fn new_unchecked(slice: &[u8]) -> &Self

Perform a raw conversion of a byte slice to Self without first performing a length check.

pub(crate) const fn as_ptr(&self) -> *const BytesRef

Get a pointer to this BytesRef.

pub const fn as_slice(&self) -> &[u8]

Borrow the inner byte slice

pub fn len(&self) -> Length

Get the Length of this BytesRef.

pub const fn is_empty(&self) -> bool

Is this BytesRef empty?

pub fn prefix(&self, length: Length) -> Result<&Self>

Get a prefix of a [crate::bytes_ref::BytesRef] of the given length.