Enum der::Tag

#[non_exhaustive]
#[repr(u8)]
pub enum Tag {

    Boolean,
    Integer,
    BitString,
    OctetString,
    Null,
    ObjectIdentifier,
    Utf8String,
    Set,
    PrintableString,
    Ia5String,
    UtcTime,
    GeneralizedTime,
    Sequence,
    ContextSpecific0,
    ContextSpecific1,
    ContextSpecific2,
    ContextSpecific3,
    ContextSpecific4,
    ContextSpecific5,
    ContextSpecific6,
    ContextSpecific7,
    ContextSpecific8,
    ContextSpecific9,
    ContextSpecific10,
    ContextSpecific11,
    ContextSpecific12,
    ContextSpecific13,
    ContextSpecific14,
    ContextSpecific15,

}

ASN.1 tags.

Tags are the leading byte of the Tag-Length-Value encoding used by ASN.1 DER and identify the type of the subsequent value.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Boolean

BOOLEAN tag.

Integer

INTEGER tag.

BitString

BIT STRING tag.

OctetString

OCTET STRING tag.

Null

NULL tag.

ObjectIdentifier

OBJECT IDENTIFIER tag.

Utf8String

UTF8String tag.

Set

SET and SET OF tag.

PrintableString

PrintableString tag.

Ia5String

IA5String tag.

UtcTime

UTCTime tag.

GeneralizedTime

GeneralizedTime tag.

Sequence

SEQUENCE tag.

Note that the universal tag number for SEQUENCE is technically 0x10 however we presently only support the constructed form, which has the 6th bit (i.e. 0x20) set.

ContextSpecific0

Context-specific tag (0) unique to a particular structure.

ContextSpecific1

Context-specific tag (1) unique to a particular structure.

ContextSpecific2

Context-specific tag (2) unique to a particular structure.

ContextSpecific3

Context-specific tag (3) unique to a particular structure.

ContextSpecific4

Context-specific tag (4) unique to a particular structure.

ContextSpecific5

Context-specific tag (5) unique to a particular structure.

ContextSpecific6

Context-specific tag (6) unique to a particular structure.

ContextSpecific7

Context-specific tag (7) unique to a particular structure.

ContextSpecific8

Context-specific tag (8) unique to a particular structure.

ContextSpecific9

Context-specific tag (9) unique to a particular structure.

ContextSpecific10

Context-specific tag (10) unique to a particular structure.

ContextSpecific11

Context-specific tag (11) unique to a particular structure.

ContextSpecific12

Context-specific tag (12) unique to a particular structure.

ContextSpecific13

Context-specific tag (13) unique to a particular structure.

ContextSpecific14

Context-specific tag (14) unique to a particular structure.

ContextSpecific15

Context-specific tag (15) unique to a particular structure.

Implementations

impl Tag

pub fn assert_eq(self, expected: Tag) -> Result<Tag>

Assert that this Tag matches the provided expected tag.

On mismatch, returns an Error with ErrorKind::UnexpectedTag.

pub fn class(self) -> Class

Get the Class that corresponds to this Tag.

pub fn context_specific(tag: u8) -> Result<Tag>

Get the Tag value that corresponds to a context-specific tag value (i.e. lower 6-bits of the tag, sans leading 10 bits)

pub fn is_context_specific(self) -> bool

Is this a context-specific tag?

pub fn type_name(self) -> &'static str

Names of ASN.1 type which corresponds to a given Tag.