der::ord

Trait OrdIsValueOrd

pub trait OrdIsValueOrd: Ord { }

Marker trait for types whose Ord impl can be used as ValueOrd.

This means the Ord impl will sort values in the same order as their DER encodings.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.