der::referenced

Trait RefToOwned

pub trait RefToOwned<'a> {
    type Owned: OwnedToRef<Borrowed<'a> = Self>
       where Self: 'a;

    // Required method
    fn ref_to_owned(&self) -> Self::Owned;
}

A trait for cloning a referenced structure and getting owned objects

This is the pendant to OwnedToRef.

This converts an object borrowing data to one that will copy the data over and own the content.

Required Associated Types

type Owned: OwnedToRef<Borrowed<'a> = Self> where Self: 'a

The resulting type after obtaining ownership.

Required Methods

fn ref_to_owned(&self) -> Self::Owned

Creates a new object taking ownership of the data