der::string

Struct StringRef

#[repr(transparent)]
pub struct StringRef(str);

String slice newtype which respects the Length::max limit.

Tuple Fields
0: str
Implementations

impl StringRef

pub const fn new(s: &str) -> Result<&Self>

Create a new StringRef, ensuring that the byte representation of the provided str value is shorter than Length::max().

pub(crate) const fn new_unchecked(s: &str) -> &Self

Perform a raw conversion of a str to Self without first performing a length check.

pub fn from_bytes(bytes: &[u8]) -> Result<&Self>

Parse a StringRef from UTF-8 encoded bytes.

pub fn as_str(&self) -> &str

Borrow the inner str.

pub fn as_bytes(&self) -> &[u8]

Borrow the inner byte slice.

pub fn len(&self) -> Length

Get the Length of this StringRef.

pub fn is_empty(&self) -> bool

Is this StringRef empty?