Struct der::IndefiniteLength

pub struct IndefiniteLength(/* private fields */);

Length type with support for indefinite lengths as used by ASN.1 BER, as described in X.690 Section 8.1.3.6:

8.1.3.6 For the indefinite form, the length octets indicate that the contents octets are terminated by end-of-contents octets (see 8.1.5), and shall consist of a single octet.

8.1.3.6.1 The single octet shall have bit 8 set to one, and bits 7 to 1 set to zero.

8.1.3.6.2 If this form of length is used, then end-of-contents octets (see 8.1.5) shall be present in the encoding following the contents octets.

Indefinite lengths are non-canonical and therefore invalid DER, however there are interoperability corner cases where we have little choice but to tolerate some BER productions where this is helpful.

Implementations

impl IndefiniteLength

pub const ZERO: Self = _

Length of 0.

pub const ONE: Self = _

Length of 1.

pub const INDEFINITE: Self = _

Indefinite length.

impl IndefiniteLength

pub fn new(length: impl Into<Length>) -> Self

Create a definite length from a type which can be converted into a Length.

pub fn is_definite(self) -> bool

Is this length definite?

pub fn is_indefinite(self) -> bool

Is this length indefinite?