der

Struct Length

pub struct Length { /* private fields */ }

ASN.1-encoded length.

Examples

use der::{Decode, Length, SliceReader};

let mut reader = SliceReader::new(&[0x82, 0xAA, 0xBB]).unwrap();
let length = Length::decode(&mut reader).expect("valid length");

assert_eq!(length, Length::new(0xAABB));

5-byte lengths are supported:

use der::{Encode, Length};
let length = Length::new(0x10000000);

assert_eq!(length.encoded_len(), Ok(Length::new(5)));

Invalid lengths produce an error:

use der::{Decode, Length, SliceReader};

let mut reader = SliceReader::new(&[0x81, 0x7F]).unwrap();

Length::decode(&mut reader).expect_err("non-canonical length should be rejected");

Implementations

impl Length

pub const ZERO: Self

Length of 0

pub const ONE: Self

Length of 1

pub const MAX: Self

Maximum length (u32::MAX).

pub const fn new(value: u32) -> Self

Create a new Length for any value which fits inside of a u16.

This function is const-safe and therefore useful for Length constants.

pub const fn is_zero(self) -> bool

Is this length equal to zero?

pub fn for_tlv(self, tag: Tag) -> Result<Self>

Get the length of DER Tag-Length-Value (TLV) encoded data if self is the length of the inner “value” portion of the message.

Errors

Returns an error if an overflow occurred computing the length.

pub fn saturating_add(self, rhs: Self) -> Self

Perform saturating addition of two lengths.

pub fn saturating_sub(self, rhs: Self) -> Self

Perform saturating subtraction of two lengths.