der
Struct PemWriter

pub struct PemWriter<'w>(/* private fields */);

Writer type which outputs PEM-encoded data.

Implementations

impl<'w> PemWriter<'w>

pub fn new( type_label: &'static str, line_ending: LineEnding, out: &'w mut [u8], ) -> Result<Self>

Create a new PEM writer which outputs into the provided buffer.

Uses the default 64-character line wrapping.

pub fn type_label(&self) -> &'static str

Get the PEM label which will be used in the encapsulation boundaries for this document.

pub fn finish(self) -> Result<usize>

Finish encoding PEM, writing the post-encapsulation boundary.

On success, returns the total number of bytes written to the output buffer.