Struct der::SliceReader

pub struct SliceReader<'a> { /* private fields */ }

Reader which consumes an input byte slice.

Implementations

impl<'a> SliceReader<'a>

pub fn new(bytes: &'a [u8]) -> Result<Self>

Create a new slice reader for the given byte slice.

pub fn error(&mut self, kind: ErrorKind) -> Error

Return an error with the given ErrorKind, annotating it with context about where the error occurred.

pub fn value_error(&mut self, tag: Tag) -> Error

Return an error for an invalid value with the given tag.

pub fn is_failed(&self) -> bool

Did the decoding operation fail due to an error?