Struct der::SliceWriter

pub struct SliceWriter<'a> { /* private fields */ }

Writer which encodes DER into a mutable output byte slice.

Implementations

impl<'a> SliceWriter<'a>

pub fn new(bytes: &'a mut [u8]) -> Self

Create a new encoder with the given byte slice as a backing buffer.

pub fn encode<T: Encode>(&mut self, encodable: &T) -> Result<()>

Encode a value which impls the Encode trait.

pub fn error<T>(&mut self, kind: ErrorKind) -> Result<T>

Return an error with the given ErrorKind, annotating it with context about where the error occurred.

pub fn is_failed(&self) -> bool

Did the decoding operation fail due to an error?

pub fn finish(self) -> Result<&'a [u8]>

Finish encoding to the buffer, returning a slice containing the data written to the buffer.

pub fn context_specific<T>(
&mut self,
tag_number: TagNumber,
tag_mode: TagMode,
value: &T
) -> Result<()>where
T: EncodeValue + Tagged,

Encode a CONTEXT-SPECIFIC field with the provided tag number and mode.

pub fn sequence<F>(&mut self, length: Length, f: F) -> Result<()>where
F: FnOnce(&mut SliceWriter<'_>) -> Result<()>,

Encode an ASN.1 SEQUENCE of the given length.

Spawns a nested slice writer which is expected to be exactly the specified length upon completion.