Trait der::Choice

pub trait Choice<'a>: Decodable<'a> + Encodable {
    fn can_decode(tag: Tag) -> bool;
}

ASN.1 CHOICE denotes a union of one or more possible alternatives.

The types MUST have distinct tags.

This crate models choice as a trait, with a blanket impl for all types which impl Decodable + Encodable + Tagged (i.e. they are modeled as a CHOICE with only one possible variant)

Required methods

fn can_decode(tag: Tag) -> bool

Is the provided Tag decodable as a variant of this CHOICE?