Trait der::DecodeOwned

pub trait DecodeOwned: for<'a> Decode<'a> { }

Marker trait for data structures that can be decoded from DER without borrowing any data from the decoder.

This is primarily useful for trait bounds on functions which require that no data is borrowed from the decoder, for example a PEM decoder which needs to first decode data from Base64.

This trait is inspired by the DeserializeOwned trait from serde.