der

Trait DecodeValue

pub trait DecodeValue<'a>: Sized {
    type Error: Error + From<Error> + 'static;

    // Required method
    fn decode_value<R: Reader<'a>>(
        reader: &mut R,
        header: Header,
    ) -> Result<Self, Self::Error>;
}

DecodeValue trait parses the value part of a Tag-Length-Value object, sans the Tag and Length.

As opposed to Decode, implementer is expected to read the inner content only, without the Header, which was decoded beforehand.

Example

use der::{Decode, DecodeValue, ErrorKind, FixedTag, Header, Reader, Tag};

/// 1-byte month
struct MyByteMonth(u8);

impl<'a> DecodeValue<'a> for MyByteMonth {
    type Error = der::Error;

    fn decode_value<R: Reader<'a>>(reader: &mut R, header: Header) -> der::Result<Self> {
        let month = reader.read_byte()?;
         
        if (0..12).contains(&month) {
            Ok(Self(month))
        } else {
            Err(reader.error(ErrorKind::DateTime))
        }
    }
}

impl FixedTag for MyByteMonth {
    const TAG: Tag = Tag::OctetString;
}

let month = MyByteMonth::from_der(b"\x04\x01\x09").expect("month to decode");

assert_eq!(month.0, 9);

Required Associated Types

type Error: Error + From<Error> + 'static

Type returned in the event of a decoding error.

Required Methods

fn decode_value<R: Reader<'a>>( reader: &mut R, header: Header, ) -> Result<Self, Self::Error>

Attempt to decode this value using the provided Reader.

Errors

Returns Self::Error in the event a decoding error occurred.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.