der

Trait DerOrd

pub trait DerOrd {
    // Required method
    fn der_cmp(&self, other: &Self) -> Result<Ordering>;
}

DER ordering trait.

Compares the ordering of two values based on their ASN.1 DER serializations.

This is used by the DER encoding for SET OF in order to establish an ordering for the elements of sets.

Required Methods

fn der_cmp(&self, other: &Self) -> Result<Ordering>

Return an Ordering between self and other when serialized as ASN.1 DER.

Errors

If an encoding error occurred during the comparison.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.