der

Trait Encodable

pub trait Encodable {
    // Required methods
    fn encoded_len(&self) -> Result<Length>;
    fn encode(&self, encoder: &mut Encoder<'_>) -> Result<()>;

    // Provided methods
    fn encode_to_slice<'a>(&self, buf: &'a mut [u8]) -> Result<&'a [u8]> { ... }
    fn encode_to_vec(&self, buf: &mut Vec<u8>) -> Result<Length> { ... }
    fn to_vec(&self) -> Result<Vec<u8>> { ... }
}

Encoding trait.

Required Methods

fn encoded_len(&self) -> Result<Length>

Compute the length of this value in bytes when encoded as ASN.1 DER.

fn encode(&self, encoder: &mut Encoder<'_>) -> Result<()>

Encode this value as ASN.1 DER using the provided Encoder.

Provided Methods

fn encode_to_slice<'a>(&self, buf: &'a mut [u8]) -> Result<&'a [u8]>

Encode this value to the provided byte slice, returning a sub-slice containing the encoded message.

fn encode_to_vec(&self, buf: &mut Vec<u8>) -> Result<Length>

Encode this message as ASN.1 DER, appending it to the provided byte vector.

fn to_vec(&self) -> Result<Vec<u8>>

Serialize this message as a byte vector.