der

Trait EncodeValue

pub trait EncodeValue {
    // Required methods
    fn value_len(&self) -> Result<Length>;
    fn encode_value(&self, writer: &mut impl Writer) -> Result<()>;

    // Provided method
    fn header(&self) -> Result<Header>
       where Self: Tagged { ... }
}

Encode the value part of a Tag-Length-Value encoded field, sans the Tag and Length.

As opposed to Encode, implementer is expected to write the inner content only, without the Header.

When EncodeValue is paired with FixedTag, it produces a complete TLV ASN.1 DER encoding as Encode trait.

Example

use der::{Encode, EncodeValue, ErrorKind, FixedTag, Length, Tag, Writer};

/// 1-byte month
struct MyByteMonth(u8);

impl EncodeValue for MyByteMonth {

    fn value_len(&self) -> der::Result<Length> {
        Ok(Length::new(1))
    }

    fn encode_value(&self, writer: &mut impl Writer) -> der::Result<()> {
        writer.write_byte(self.0)?;
        Ok(())
    }
}

impl FixedTag for MyByteMonth {
    const TAG: Tag = Tag::OctetString;
}

let month = MyByteMonth(9);
let mut buf = [0u8; 16];
let month_der = month.encode_to_slice(&mut buf).expect("month to encode");

assert_eq!(month_der, b"\x04\x01\x09");

Required Methods

fn value_len(&self) -> Result<Length>

Compute the length of this value (sans Tag+Length header) when encoded as ASN.1 DER.

Errors

Returns an error if the value length could not be computed (e.g. overflow).

fn encode_value(&self, writer: &mut impl Writer) -> Result<()>

Encode value (sans Tag+Length header) as ASN.1 DER using the provided Writer.

Errors

In the event an encoding error occurred.

Provided Methods

fn header(&self) -> Result<Header>
where Self: Tagged,

Get the Header used to encode this value.

Errors

Returns an error if the header could not be computed.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.