der

Module asn1

Module containing all of the various ASN.1 built-in types supported by this library.

Structs

Any

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

AnyRef

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

BitString

Owned form of ASN.1 BIT STRING type.

BitStringIter

Iterator over the bits of a BitString.

BitStringRef

ASN.1 BIT STRING type.

BmpString

ASN.1 BMPString type.

ContextSpecific

Context-specific field which wraps an owned inner value.

ContextSpecificRef

Context-specific field reference.

GeneralizedTime

ASN.1 GeneralizedTime type.

Ia5String

ASN.1 IA5String type.

Ia5StringRef

ASN.1 IA5String type.

Int

Signed arbitrary precision ASN.1 INTEGER type.

IntRef

Signed arbitrary precision ASN.1 INTEGER reference type.

Null

ASN.1 NULL type.

ObjectIdentifier

Object identifier (OID).

OctetString

ASN.1 OCTET STRING type: owned form..

OctetStringRef

ASN.1 OCTET STRING type: borrowed form.

PrintableString

ASN.1 PrintableString type.

PrintableStringRef

ASN.1 PrintableString type.

SequenceOf

ASN.1 SEQUENCE OF backed by an array.

SequenceOfIter

Iterator over the elements of an SequenceOf.

SequenceRef

The SequenceRef type provides raw access to the octets which comprise a DER-encoded SEQUENCE.

SetOf

ASN.1 SET OF backed by an array.

SetOfIter

Iterator over the elements of an SetOf.

SetOfVec

ASN.1 SET OF backed by a Vec.

TeletexString

ASN.1 TeletexString type.

TeletexStringRef

ASN.1 TeletexString type.

Uint

Unsigned arbitrary precision ASN.1 INTEGER type.

UintRef

Unsigned arbitrary precision ASN.1 INTEGER reference type.

UtcTime

ASN.1 UTCTime type.

Utf8StringRef

ASN.1 UTF8String type.

VideotexStringRef

ASN.1 VideotexString type.

Traits

Choice

ASN.1 CHOICE denotes a union of one or more possible alternatives.

Sequence

Marker trait for ASN.1 SEQUENCEs.