der

Module asn1

Module containing all of the various ASN.1 built-in types supported by this library.

Modules

any 🔒

ASN.1 ANY type.

application 🔒

Application class field.

bit_string 🔒

ASN.1 BIT STRING support.

bmp_string 🔒

ASN.1 BMPString support.

boolean 🔒

ASN.1 BOOLEAN support.

choice 🔒

ASN.1 CHOICE support.

context_specific 🔒

Context-specific field.

general_string 🔒

generalized_time 🔒

ASN.1 GeneralizedTime support.

ia5_string 🔒

ASN.1 IA5String support.

integer 🔒

ASN.1 INTEGER support.

internal_macros 🔒

null 🔒

ASN.1 NULL support.

octet_string 🔒

ASN.1 OCTET STRING support.

oid 🔒

ASN.1 OBJECT IDENTIFIER

optional 🔒

ASN.1 OPTIONAL as mapped to Rust’s Option type

printable_string 🔒

ASN.1 PrintableString support.

private 🔒

Private class field.

sequence 🔒

The Sequence trait simplifies writing decoders/encoders which map ASN.1 SEQUENCEs to Rust structs.

sequence_of 🔒

ASN.1 SEQUENCE OF support.

set_of 🔒

ASN.1 SET OF support.

teletex_string 🔒

ASN.1 TeletexString support.

utc_time 🔒

ASN.1 UTCTime support.

utf8_string 🔒

ASN.1 UTF8String support.

videotex_string 🔒

ASN.1 VideotexString support.

Structs

Any

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

AnyRef

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

Application

APPLICATION field which wraps an owned inner value.

ApplicationRef

APPLICATION field reference.

BitString

Owned form of ASN.1 BIT STRING type.

BitStringIter

Iterator over the bits of a BitString.

BitStringRef

ASN.1 BIT STRING type.

BmpString

ASN.1 BMPString type.

ContextSpecific

CONTEXT-SPECIFIC field which wraps an owned inner value.

ContextSpecificRef

CONTEXT-SPECIFIC field reference.

GeneralStringRef

This is currently &[u8] internally, as GeneralString is not fully implemented yet

GeneralizedTime

ASN.1 GeneralizedTime type.

Ia5String

ASN.1 IA5String type.

Ia5StringRef

ASN.1 IA5String type.

Int

Signed arbitrary precision ASN.1 INTEGER type.

IntRef

Signed arbitrary precision ASN.1 INTEGER reference type.

Null

ASN.1 NULL type.

ObjectIdentifier

Object identifier (OID).

OctetString

ASN.1 OCTET STRING type: owned form.

OctetStringRef

ASN.1 OCTET STRING type: borrowed form.

PrintableString

ASN.1 PrintableString type.

PrintableStringRef

ASN.1 PrintableString type.

Private

PRIVATE field which wraps an owned inner value.

PrivateRef

PRIVATE field reference.

SequenceRef

The SequenceRef type provides raw access to the octets which comprise a DER-encoded SEQUENCE.

SetOfIter

Iterator over the elements of an [SetOf].

SetOfRef

ASN.1 SET OF with a reference to an array.

SetOfVec

ASN.1 SET OF backed by a Vec.

TeletexString

ASN.1 TeletexString type.

TeletexStringRef

ASN.1 TeletexString type.

Uint

Unsigned arbitrary precision ASN.1 INTEGER type.

UintRef

Unsigned arbitrary precision ASN.1 INTEGER reference type.

UtcTime

ASN.1 UTCTime type.

Utf8StringRef

ASN.1 UTF8String type.

VideotexStringRef

ASN.1 VideotexString type.

Traits

Choice

ASN.1 CHOICE denotes a union of one or more possible alternatives.

Sequence

Marker trait for ASN.1 SEQUENCEs.