Module der::asn1

ASN.1 built-in types.

Structs

Any
ASN.1 ANY: represents any explicitly tagged ASN.1 value.

BitString
ASN.1 BIT STRING type.

ContextSpecific
Context-specific field.

GeneralizedTime
ASN.1 GeneralizedTime type.

Ia5String
ASN.1 IA5String type.

Null
ASN.1 NULL type.

OctetString
ASN.1 OCTET STRING type.

PrintableString
ASN.1 PrintableString type.

Sequence
ASN.1 SEQUENCE type.

SequenceIter
ASN.1 SEQUENCE iterator for Sequence containing homogeneously typed values.

SetOfRef
ASN.1 SET OF backed by a byte slice containing serialized DER.

SetOfRefIter
Iterator over the elements of an SetOfRef.

UIntBytes
“Big” unsigned ASN.1 INTEGER type.

UtcTime
ASN.1 UTCTime type.

Utf8String
ASN.1 UTF8String type.

Traits

SetOf
ASN.1 SET OF denotes a collection of zero or more occurrences of a given type.