Struct der::asn1::Any

pub struct Any { /* private fields */ }

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

This type provides the same functionality as AnyRef but owns the backing data.

Implementations

impl Any

pub fn new(tag: Tag, bytes: impl Into<Box<[u8]>>) -> Result<Self>

Create a new Any from the provided Tag and DER bytes.

pub fn value(&self) -> &[u8]

Allow access to value

pub fn decode_as<'a, T>(&'a self) -> Result<T>
where T: Choice<'a> + DecodeValue<'a>,

Attempt to decode this Any type into the inner value.

pub fn encode_from<T>(msg: &T) -> Result<Self>
where T: Tagged + EncodeValue,

Encode the provided type as an Any value.

pub fn sequence<'a, F, T>(&'a self, f: F) -> Result<T>
where F: FnOnce(&mut SliceReader<'a>) -> Result<T>,

Attempt to decode this value an ASN.1 SEQUENCE, creating a new nested reader and calling the provided argument with it.

pub fn null() -> Self

Any representation of the ASN.1 NULL type.

impl Any

pub fn is_null(&self) -> bool

Is this value an ASN.1 NULL value?