Struct der::asn1::Any

pub struct Any<'a> { /* private fields */ }

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

Technically ANY hasn’t been a recommended part of ASN.1 since the X.209 revision from 1988. It was deprecated and replaced by Information Object Classes in X.680 in 1994, and X.690 no longer refers to it whatsoever.

Nevertheless, this crate defines an Any type as it remains a familiar and useful concept which is still extensively used in things like PKI-related RFCs.

Implementations

impl<'a> Any<'a>

pub fn new(tag: Tag, bytes: &'a [u8]) -> Result<Self>

Create a new Any from the provided Tag and byte slice.

pub fn tag(self) -> Tag

Get the tag for this Any type.

pub fn len(self) -> Length

Get the Length of this Any type’s value.

pub fn is_empty(self) -> bool

Is the body of this Any type empty?

pub fn is_null(self) -> bool

Is this value an ASN.1 NULL value?

pub fn as_bytes(self) -> &'a [u8]

Get the raw value for this Any type as a byte slice.

pub fn bit_string(self) -> Result<BitString<'a>>

Attempt to decode an ASN.1 BIT STRING.

pub fn context_specific(self) -> Result<ContextSpecific<'a>>

Attempt to decode an ASN.1 CONTEXT-SPECIFIC field.

pub fn generalized_time(self) -> Result<GeneralizedTime>

Attempt to decode an ASN.1 GeneralizedTime.

pub fn ia5_string(self) -> Result<Ia5String<'a>>

Attempt to decode an ASN.1 IA5String.

pub fn octet_string(self) -> Result<OctetString<'a>>

Attempt to decode an ASN.1 OCTET STRING.

pub fn optional<T>(self) -> Result<Option<T>> where
T: Choice<'a> + TryFrom<Self, Error = Error>,

Attempt to decode an ASN.1 OPTIONAL value.

pub fn printable_string(self) -> Result<PrintableString<'a>>

Attempt to decode an ASN.1 PrintableString.

pub fn sequence<F, T>(self, f: F) -> Result<T> where
F: FnOnce(&mut Decoder<'a>) -> Result<T>,

Attempt to decode this value an ASN.1 SEQUENCE, creating a new nested Decoder and calling the provided argument with it.

pub fn utc_time(self) -> Result<UtcTime>

Attempt to decode an ASN.1 UTCTime.

pub fn utf8_string(self) -> Result<Utf8String<'a>>

Attempt to decode an ASN.1 UTF8String.