Struct der::asn1::AnyRef

pub struct AnyRef<'a> { /* private fields */ }

ASN.1 ANY: represents any explicitly tagged ASN.1 value.

This is a zero-copy reference type which borrows from the input data.

Technically ANY hasn’t been a recommended part of ASN.1 since the X.209 revision from 1988. It was deprecated and replaced by Information Object Classes in X.680 in 1994, and X.690 no longer refers to it whatsoever.

Nevertheless, this crate defines an ANY type as it remains a familiar and useful concept which is still extensively used in things like PKI-related RFCs.

Implementations

impl<'a> AnyRef<'a>

pub const NULL: Self = _

AnyRef representation of the ASN.1 NULL type.

pub fn new(tag: Tag, bytes: &'a [u8]) -> Result<Self>

Create a new AnyRef from the provided Tag and DER bytes.

pub fn value(self) -> &'a [u8]

Get the raw value for this AnyRef type as a byte slice.

pub fn decode_into<T>(self) -> Result<T>where
T: DecodeValue<'a> + FixedTag,

Attempt to decode this AnyRef type into the inner value.

pub fn is_null(self) -> bool

Is this value an ASN.1 NULL value?

pub fn bit_string(self) -> Result<BitStringRef<'a>>

Attempt to decode an ASN.1 BIT STRING.

pub fn context_specific<T>(self) -> Result<ContextSpecific<T>>where
T: Decode<'a>,

Attempt to decode an ASN.1 CONTEXT-SPECIFIC field.

pub fn generalized_time(self) -> Result<GeneralizedTime>

Attempt to decode an ASN.1 GeneralizedTime.

pub fn ia5_string(self) -> Result<Ia5StringRef<'a>>

Attempt to decode an ASN.1 IA5String.

pub fn octet_string(self) -> Result<OctetStringRef<'a>>

Attempt to decode an ASN.1 OCTET STRING.

pub fn oid(self) -> Result<ObjectIdentifier>

Attempt to decode an ASN.1 OBJECT IDENTIFIER.

pub fn optional<T>(self) -> Result<Option<T>>where
T: Choice<'a> + TryFrom<Self, Error = Error>,

Attempt to decode an ASN.1 OPTIONAL value.

pub fn printable_string(self) -> Result<PrintableStringRef<'a>>

Attempt to decode an ASN.1 PrintableString.

pub fn teletex_string(self) -> Result<TeletexStringRef<'a>>

Attempt to decode an ASN.1 TeletexString.

pub fn videotex_string(self) -> Result<VideotexStringRef<'a>>

Attempt to decode an ASN.1 VideotexString.

pub fn sequence<F, T>(self, f: F) -> Result<T>where
F: FnOnce(&mut SliceReader<'a>) -> Result<T>,

Attempt to decode this value an ASN.1 SEQUENCE, creating a new nested reader and calling the provided argument with it.

pub fn utc_time(self) -> Result<UtcTime>

Attempt to decode an ASN.1 UTCTime.

pub fn utf8_string(self) -> Result<Utf8StringRef<'a>>

Attempt to decode an ASN.1 UTF8String.