Struct der::asn1::BitString

pub struct BitString<'a> { /* fields omitted */ }

ASN.1 BIT STRING type.

This type contains a sequence of any number of bits, modeled internally as a sequence of bytes with a known number of “unused bits”.

Implementations

impl<'a> BitString<'a>

pub const MAX_UNUSED_BITS: u8

Maximum number of unused bits allowed.

pub fn new(unused_bits: u8, bytes: &'a [u8]) -> Result<Self>

Create a new ASN.1 BIT STRING from a byte slice.

Accepts an optional number of “unused bits” (0-7) which are omitted from the final octet. This number is 0 if the value is octet-aligned.

pub fn from_bytes(bytes: &'a [u8]) -> Result<Self>

Create a new ASN.1 BIT STRING from the given bytes.

The “unused bits” are set to 0.

pub fn unused_bits(&self) -> u8

Get the number of unused bits in this byte slice.

pub fn has_unused_bits(&self) -> bool

Is the number of unused bits a value other than 0?

pub fn bit_len(&self) -> usize

Get the length of this BIT STRING in bits.

pub fn byte_len(&self) -> Length

Get the number of bytes/octets needed to represent this BIT STRING when serialized in an octet-aligned manner.

pub fn is_empty(&self) -> bool

Is the inner byte slice empty?

pub fn as_bytes(&self) -> Option<&'a [u8]>

Borrow the inner byte slice.

Returns None if the number of unused bits is not equal to zero, i.e. if the BIT STRING is not octet aligned.

Use BitString::raw_bytes to obtain access to the raw value regardless of the presence of unused bits.

pub fn raw_bytes(&self) -> &'a [u8]

Borrow the raw bytes of this BIT STRING.

Note that the byte string may contain extra unused bits in the final octet. If the number of unused bits is expected to be 0, the BitString::as_bytes function can be used instead.

pub fn bits(self) -> BitStringIter<'a>

Iterator over the bits of this BIT STRING.