Struct der::asn1::SequenceRef

pub struct SequenceRef<'a> { /* private fields */ }

The SequenceRef type provides raw access to the octets which comprise a DER-encoded SEQUENCE.

This is a zero-copy reference type which borrows from the input data.