der::asn1
Struct UtcTime

pub struct UtcTime(/* private fields */);

ASN.1 UTCTime type.

This type implements the validity requirements specified in RFC 5280 Section 4.1.2.5.1, namely:

For the purposes of this profile, UTCTime values MUST be expressed in Greenwich Mean Time (Zulu) and MUST include seconds (i.e., times are YYMMDDHHMMSSZ), even where the number of seconds is zero. Conforming systems MUST interpret the year field (YY) as follows:

Where YY is greater than or equal to 50, the year SHALL be interpreted as 19YY; and

Where YY is less than 50, the year SHALL be interpreted as 20YY.

Note: Due to common operations working on UNIX_EPOCH UtcTimes are only supported for the years 1970-2049.

Implementations

impl UtcTime

pub const LENGTH: usize = 13usize

Length of an RFC 5280-flavored ASN.1 DER-encoded UtcTime.

pub const MAX_YEAR: u16 = 2_049u16

Maximum year that can be represented as a UTCTime.

pub fn from_date_time(datetime: DateTime) -> Result<Self>

Create a UtcTime from a DateTime.

pub fn to_date_time(&self) -> DateTime

Convert this UtcTime into a DateTime.

pub fn from_unix_duration(unix_duration: Duration) -> Result<Self>

Create a new UtcTime given a Duration since UNIX_EPOCH (a.k.a. “Unix time”)

pub fn to_unix_duration(&self) -> Duration

Get the duration of this timestamp since UNIX_EPOCH.

pub fn from_system_time(time: SystemTime) -> Result<Self>

Instantiate from SystemTime.

pub fn to_system_time(&self) -> SystemTime

Convert to SystemTime.