Enum der::ErrorKind

#[non_exhaustive]
pub enum ErrorKind {

    DuplicateField {
        tag: Tag,
    },
    Failed,
    Length {
        tag: Tag,
    },
    Noncanonical {
        tag: Tag,
    },
    MalformedOid,
    Overflow,
    Overlength,
    TrailingData {
        decoded: Length,
        remaining: Length,
    },
    Truncated,
    Underlength {
        expected: Length,
        actual: Length,
    },
    UnexpectedTag {
        expected: Option<Tag>,
        actual: Tag,
    },
    UnknownTag {
        byte: u8,
    },
    Utf8(Utf8Error),
    Value {
        tag: Tag,
    },

}

Error type.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

DuplicateField

Fields

tag: Tag

Tag of the duplicated field.

Indicates a field which is duplicated when only one is expected.

Failed

This error indicates a previous DER parsing operation resulted in an error and tainted the state of a Decoder or Encoder.

Once this occurs, the overall operation has failed and cannot be subsequently resumed.

Length

Fields

tag: Tag

Tag of the value being decoded.

Incorrect length for a given field.

Noncanonical

Fields

tag: Tag

Tag of the value which is not canonically encoded.

Message is not canonically encoded.

MalformedOid

Malformed OID

Overflow

Integer overflow occurred (library bug!).

Overlength

Message is longer than this library’s internal limits support.

TrailingData

Fields

decoded: Length

Length of the decoded data.

remaining: Length

Total length of the remaining data left in the buffer.

Undecoded trailing data at end of message.

Truncated

Unexpected end-of-message/nested field when decoding.

Underlength

Fields

expected: Length

Expected length

actual: Length

Actual length

Encoded message is shorter than the expected length.

(i.e. an Encodable impl on a particular type has a buggy encoded_len)

UnexpectedTag

Fields

expected: Option<Tag>

Tag the decoder was expecting (if there is a single such tag).

None if multiple tags are expected/allowed, but the actual tag does not match any of them.

actual: Tag

Actual tag encountered in the message.

Unexpected tag.

UnknownTag

Fields

byte: u8

Raw byte value of the tag.

Unknown/unsupported tag.

Utf8(Utf8Error)

UTF-8 errors.

Value

Fields

tag: Tag

Tag of the unexpected value.

Unexpected value.

Implementations

impl ErrorKind

pub fn at(self, position: Length) -> Error

Annotate an ErrorKind with context about where it occurred, returning an error.