Enum der::Tag

#[non_exhaustive]
#[repr(u8)]pub enum Tag {
    Boolean,
    Integer,
    BitString,
    OctetString,
    Null,
    ObjectIdentifier,
    Utf8String,
    Set,
    PrintableString,
    Ia5String,
    UtcTime,
    GeneralizedTime,
    Sequence,
    ContextSpecific0,
    ContextSpecific1,
    ContextSpecific2,
    ContextSpecific3,
}

ASN.1 tags.

Tags are the leading byte of the Tag-Length-Value encoding used by ASN.1 DER and identify the type of the subsequent value.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Boolean

BOOLEAN tag.

Integer

INTEGER tag.

BitString

BIT STRING tag.

OctetString

OCTET STRING tag.

Null

NULL tag.

ObjectIdentifier

OBJECT IDENTIFIER tag.

Utf8String

UTF8String tag.

Set

SET and SET OF tag.

PrintableString

PrintableString tag.

Ia5String

IA5String tag.

UtcTime

UTCTime tag.

GeneralizedTime

GeneralizedTime tag.

Sequence

SEQUENCE tag.

Note that the universal tag number for SEQUENCE is technically 0x10 however we presently only support the constructed form, which has the 6th bit (i.e. 0x20) set.

ContextSpecific0

Context-specific tag (0) unique to a particular structure.

ContextSpecific1

Context-specific tag (1) unique to a particular structure.

ContextSpecific2

Context-specific tag (2) unique to a particular structure.

ContextSpecific3

Context-specific tag (3) unique to a particular structure.

Implementations

impl Tag

pub fn assert_eq(self, expected: Tag) -> Result<Tag>

Assert that this Tag matches the provided expected tag.

On mismatch, returns an Error with ErrorKind::UnexpectedTag.

pub fn type_name(self) -> &'static str

Names of ASN.1 type which corresponds to a given Tag.