Enum der::Tag

#[non_exhaustive]
pub enum Tag {

    Boolean,
    Integer,
    BitString,
    OctetString,
    Null,
    ObjectIdentifier,
    Utf8String,
    Sequence,
    Set,
    PrintableString,
    Ia5String,
    UtcTime,
    GeneralizedTime,
    Application(TagNumber),
    ContextSpecific(TagNumber),
    Private(TagNumber),

}

ASN.1 tags.

Tags are the leading identifier octet of the Tag-Length-Value encoding used by ASN.1 DER and identify the type of the subsequent value.

They are described in X.690 Section 8.1.2: Identifier octets, and structured as follows:

| Class | P/C | Tag Number |

Bits 8/7: Class

Bit 6: primitive (0) or constructed (1)

Bits 5-1: tag number

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Boolean

BOOLEAN tag: 0x01

Integer

INTEGER tag: 0x02

BitString

BIT STRING tag: 0x03

OctetString

OCTET STRING tag: 0x04

Null

NULL tag: 0x05

ObjectIdentifier

OBJECT IDENTIFIER tag: 0x06

Utf8String

UTF8String tag: 0x0C

Sequence

SEQUENCE tag: 0x10

Set

SET and SET OF tag: 0x11

PrintableString

PrintableString tag: 0x13

Ia5String

IA5String tag: 0x16

UtcTime

UTCTime tag: 0x17

GeneralizedTime

GeneralizedTime tag: 0x18

Application(TagNumber)

Application tag.

ContextSpecific(TagNumber)

Context-specific tag.

Private(TagNumber)

Private tag number.

Implementations

impl Tag

pub fn assert_eq(self, expected: Tag) -> Result<Tag>

Assert that this Tag matches the provided expected tag.

On mismatch, returns an Error with ErrorKind::UnexpectedTag.

pub fn class(self) -> Class

Get the Class that corresponds to this Tag.

pub fn octet(self) -> u8

Get the octet encoding for this Tag.

pub fn non_canonical_error(self) -> Error

Create an Error for an non-canonical value with the ASN.1 type identified by this tag.

pub fn unexpected_error(self, expected: Option<Self>) -> Error

Create an Error because the current tag was unexpected, with an optional expected tag.

pub fn value_error(self) -> Error

Create an Error for an invalid value with the ASN.1 type identified by this tag.