Struct der::Decoder

pub struct Decoder<'a> { /* private fields */ }

DER decoder.

Implementations

impl<'a> Decoder<'a>

pub fn new(bytes: &'a [u8]) -> Result<Self>

Create a new decoder for the given byte slice.

pub fn decode<T: Decodable<'a>>(&mut self) -> Result<T>

Decode a value which impls the Decodable trait.

pub fn error(&mut self, kind: ErrorKind) -> Error

Return an error with the given ErrorKind, annotating it with context about where the error occurred.

pub fn value_error(&mut self, tag: Tag) -> Error

Return an error for an invalid value with the given tag.

pub fn is_failed(&self) -> bool

Did the decoding operation fail due to an error?

pub fn position(&self) -> Length

Get the position within the buffer.

pub fn peek_byte(&self) -> Option<u8>

Peek at the next byte in the decoder without modifying the cursor.

pub fn peek_tag(&self) -> Result<Tag>

Peek at the next byte in the decoder and attempt to decode it as a Tag value.

Does not modify the decoder’s state.

pub fn peek_header(&self) -> Result<Header>

Peek forward in the decoder, attempting to decode a Header from the data at the current position in the decoder.

Does not modify the decoder’s state.

pub fn finish<T>(self, value: T) -> Result<T>

Finish decoding, returning the given value if there is no remaining data, or an error otherwise

pub fn is_finished(&self) -> bool

Have we decoded all of the bytes in this Decoder?

Returns false if we’re not finished decoding or if a fatal error has occurred.

pub fn any(&mut self) -> Result<Any<'a>>

Attempt to decode an ASN.1 ANY value.

pub fn any_optional(&mut self) -> Result<Option<Any<'a>>>

Attempt to decode an OPTIONAL ASN.1 ANY value.

pub fn int8(&mut self) -> Result<i8>

Attempt to decode ASN.1 INTEGER as i8

pub fn int16(&mut self) -> Result<i16>

Attempt to decode ASN.1 INTEGER as i16

pub fn uint8(&mut self) -> Result<u8>

Attempt to decode unsigned ASN.1 INTEGER as u8

pub fn uint16(&mut self) -> Result<u16>

Attempt to decode unsigned ASN.1 INTEGER as u16

pub fn bit_string(&mut self) -> Result<BitString<'a>>

Attempt to decode an ASN.1 BIT STRING.

pub fn context_specific<T>(
&mut self,
tag_number: TagNumber,
tag_mode: TagMode
) -> Result<Option<T>>where
T: DecodeValue<'a> + FixedTag,

Attempt to decode an ASN.1 CONTEXT-SPECIFIC field with the provided TagNumber.

pub fn generalized_time(&mut self) -> Result<GeneralizedTime>

Attempt to decode an ASN.1 GeneralizedTime.

pub fn ia5_string(&mut self) -> Result<Ia5String<'a>>

Attempt to decode an ASN.1 IA5String.

pub fn null(&mut self) -> Result<Null>

Attempt to decode an ASN.1 NULL value.

pub fn octet_string(&mut self) -> Result<OctetString<'a>>

Attempt to decode an ASN.1 OCTET STRING.

pub fn oid(&mut self) -> Result<ObjectIdentifier>

Attempt to decode an ASN.1 OBJECT IDENTIFIER.

pub fn optional<T: Choice<'a>>(&mut self) -> Result<Option<T>>

Attempt to decode an ASN.1 OPTIONAL value.

pub fn printable_string(&mut self) -> Result<PrintableString<'a>>

Attempt to decode an ASN.1 PrintableString.

pub fn utc_time(&mut self) -> Result<UtcTime>

Attempt to decode an ASN.1 UTCTime.

pub fn utf8_string(&mut self) -> Result<Utf8String<'a>>

Attempt to decode an ASN.1 UTF8String.

pub fn sequence<F, T>(&mut self, f: F) -> Result<T>where
F: FnOnce(&mut Decoder<'a>) -> Result<T>,

Attempt to decode an ASN.1 SEQUENCE, creating a new nested Decoder and calling the provided argument with it.