Struct der::Header

pub struct Header {
    pub tag: Tag,
    pub length: Length,
}

ASN.1 DER headers: tag + length component of TLV-encoded values

Fields
tag: Tag

Tag representing the type of the encoded value

length: Length

Length of the encoded value

Implementations

impl Header

pub fn new(tag: Tag, length: impl TryInto<Length>) -> Result<Self>

Create a new Header from a Tag and a specified length.

Returns an error if the length exceeds the limits of Length.