Struct der::Length

pub struct Length(_);

ASN.1-encoded length.

Maximum length is defined by the Length::MAX constant (256 MiB).

Implementations

impl Length

pub const ZERO: Self = Self(0)

Length of 0

pub const ONE: Self = Self(1)

Length of 1

pub const MAX: Self = Self(MAX_U32)

Maximum length currently supported: 256 MiB

pub const fn new(value: u16) -> Self

Create a new Length for any value which fits inside of a u16.

This function is const-safe and therefore useful for Length constants.

pub fn for_tlv(self) -> Result<Self>

Get the length of DER Tag-Length-Value (TLV) encoded data if self is the length of the inner “value” portion of the message.