Struct der::SecretDocument

pub struct SecretDocument(/* private fields */);

Secret Document type.

Useful for formats which represent potentially secret data, such as cryptographic keys.

This type provides additional hardening such as ensuring that the contents are zeroized-on-drop, and also using more restrictive file permissions when writing files to disk.

Implementations

impl SecretDocument

pub fn as_bytes(&self) -> &[u8]

Borrow the inner serialized bytes of this document.

pub fn to_bytes(&self) -> Zeroizing<Vec<u8>>

Return an allocated ASN.1 DER serialization as a byte vector.

pub fn len(&self) -> Length

Get the length of the encoded ASN.1 DER in bytes.

pub fn decode_msg<'a, T: Decode<'a>>(&'a self) -> Result<T>

Try to decode the inner ASN.1 DER message as the given type.

pub fn encode_msg<T: Encode>(msg: &T) -> Result<Self>

Encode the provided type as ASN.1 DER.

pub fn read_der_file(path: impl AsRef<Path>) -> Result<Self>

Read ASN.1 DER document from a file.

pub fn write_der_file(&self, path: impl AsRef<Path>) -> Result<()>

Write ASN.1 DER document to a file.