Trait der::Decodable

pub trait Decodable<'a>: Sized {
    fn decode(decoder: &mut Decoder<'a>) -> Result<Self>;

    fn from_der(bytes: &'a [u8]) -> Result<Self> { ... }
}

Decoding trait.

This trait provides the core abstraction upon which all decoding operations are based.

Blanket impl for TryFrom<Any>

In almost all cases you do not need to impl this trait yourself, but rather can instead impl TryFrom<Any<'a>, Error = Error> and receive a blanket impl of this trait.

Required methods

fn decode(decoder: &mut Decoder<'a>) -> Result<Self>

Attempt to decode this message using the provided decoder.

Loading content...
Provided methods

fn from_der(bytes: &'a [u8]) -> Result<Self>

Parse Self from the provided DER-encoded byte slice.

Loading content...