Trait der::Decode

pub trait Decode<'a>: Sized {
    // Required method
    fn decode<R: Reader<'a>>(decoder: &mut R) -> Result<Self>;

    // Provided method
    fn from_der(bytes: &'a [u8]) -> Result<Self> { ... }
}

Decoding trait.

This trait provides the core abstraction upon which all decoding operations are based.

Required Methods

fn decode<R: Reader<'a>>(decoder: &mut R) -> Result<Self>

Attempt to decode this message using the provided decoder.

Provided Methods

fn from_der(bytes: &'a [u8]) -> Result<Self>

Parse Self from the provided DER-encoded byte slice.

Object Safety

This trait is not object safe.