der

Trait DecodeValue

pub trait DecodeValue<'a>: Sized {
    type Error: From<Error> + 'static;

    // Required method
    fn decode_value<R: Reader<'a>>(
        reader: &mut R,
        header: Header,
    ) -> Result<Self, Self::Error>;
}

Decode the value part of a Tag-Length-Value encoded field, sans the Tag and Length.

Required Associated Types

type Error: From<Error> + 'static

Type returned in the event of a decoding error.

Required Methods

fn decode_value<R: Reader<'a>>( reader: &mut R, header: Header, ) -> Result<Self, Self::Error>

Attempt to decode this message using the provided Reader.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.