Trait der::DerOrd

pub trait DerOrd {
    // Required method
    fn der_cmp(&self, other: &Self) -> Result<Ordering>;
}

DER ordering trait.

Compares the ordering of two values based on their ASN.1 DER serializations.

This is used by the DER encoding for SET OF in order to establish an ordering for the elements of sets.

Required Methods

fn der_cmp(&self, other: &Self) -> Result<Ordering>

Return an Ordering between self and other when serialized as ASN.1 DER.

Object Safety

This trait is not object safe.