Trait der::OrdIsValueOrd

pub trait OrdIsValueOrd: Ord { }

Marker trait for types whose Ord impl can be used as ValueOrd.

This means the Ord impl will sort values in the same order as their DER encodings.