Trait der::SetOf

pub trait SetOf<'a, 'b, T>: Decodable<'a> + Encodable where
    T: Clone + Decodable<'a> + Encodable + Ord, {
    type Iter: Iterator<Item = T>;
    fn elements(&'b self) -> Self::Iter;
}

ASN.1 SET OF denotes a collection of zero or more occurrences of a given type.

When encoded as DER, SET OF is lexicographically ordered. To implement that requirement, types T which are elements of SetOf MUST provide an impl of Ord which ensures that the corresponding DER encodings of a given type are ordered.

Associated Types

type Iter: Iterator<Item = T>

Iterator over the elements of the set.

The iterator type MUST maintain the invariant that messages are lexicographically ordered.

See toplevel documentation about Ord trait requirements for more information.

Loading content...
Required methods

fn elements(&'b self) -> Self::Iter

Iterate over the elements of the set.

Loading content...